+++ erdpy_network_providers/api_network_provider_test.py
@@ -1,10 +1,10 @@
 import pytest
 from erdpy_core import Address
 
-from erdpy_network.api_network_provider import ApiNetworkProvider
-from erdpy_network.errors import GenericError
-from erdpy_network.proxy_network_provider import ContractQuery
-from erdpy_network.interface import IPagination
+from erdpy_network_providers.api_network_provider import ApiNetworkProvider
+from erdpy_network_providers.errors import GenericError
+from erdpy_network_providers.proxy_network_provider import ContractQuery
+from erdpy_network_providers.interface import IPagination
 
 
 class Pagination(IPagination):
