+++ erdpy_network_providers/contract_results.py
@@ -3,9 +3,9 @@ from typing import Any, Dict, List
 
 from erdpy_core import Address
 
-from erdpy_network.interface import IAddress
-from erdpy_network.resources import EmptyAddress
-from erdpy_network.transaction_logs import TransactionLogs
+from erdpy_network_providers.interface import IAddress
+from erdpy_network_providers.resources import EmptyAddress
+from erdpy_network_providers.transaction_logs import TransactionLogs
 
 
 class ContractResults:
