+++ erdpy_network_providers/proxy_network_provider.py
@@ -3,22 +3,22 @@ from typing import Any, Dict, List, Optional, Sequence, Tuple, Union
 import requests
 from requests.auth import AuthBase
 
-from erdpy_network.accounts import AccountOnNetwork
-from erdpy_network.constants import (DEFAULT_ADDRESS_HRP,
+from erdpy_network_providers.accounts import AccountOnNetwork
+from erdpy_network_providers.constants import (DEFAULT_ADDRESS_HRP,
                                      ESDT_CONTRACT_ADDRESS, METACHAIN_ID)
-from erdpy_network.contract_query_requests import ContractQueryRequest
-from erdpy_network.contract_query_response import ContractQueryResponse
-from erdpy_network.errors import GenericError
-from erdpy_network.interface import IAddress, IContractQuery, ITransaction
-from erdpy_network.network_config import NetworkConfig
-from erdpy_network.network_status import NetworkStatus
-from erdpy_network.resources import GenericResponse, SimulateResponse
-from erdpy_network.token_definitions import (
+from erdpy_network_providers.contract_query_requests import ContractQueryRequest
+from erdpy_network_providers.contract_query_response import ContractQueryResponse
+from erdpy_network_providers.errors import GenericError
+from erdpy_network_providers.interface import IAddress, IContractQuery, ITransaction
+from erdpy_network_providers.network_config import NetworkConfig
+from erdpy_network_providers.network_status import NetworkStatus
+from erdpy_network_providers.resources import GenericResponse, SimulateResponse
+from erdpy_network_providers.token_definitions import (
     DefinitionOfFungibleTokenOnNetwork, DefinitionOfTokenCollectionOnNetwork)
-from erdpy_network.tokens import (FungibleTokenOfAccountOnNetwork,
+from erdpy_network_providers.tokens import (FungibleTokenOfAccountOnNetwork,
                                   NonFungibleTokenOfAccountOnNetwork)
-from erdpy_network.transaction_status import TransactionStatus
-from erdpy_network.transactions import TransactionOnNetwork
+from erdpy_network_providers.transaction_status import TransactionStatus
+from erdpy_network_providers.transactions import TransactionOnNetwork
 
 
 class ProxyNetworkProvider:
