+++ erdpy_network_providers/transaction_logs.py
@@ -2,9 +2,9 @@ from typing import Any, Callable, Dict, List, Optional, Union
 
 from erdpy_core import Address
 
-from erdpy_network.interface import IAddress
-from erdpy_network.resources import EmptyAddress
-from erdpy_network.transaction_events import TransactionEvent
+from erdpy_network_providers.interface import IAddress
+from erdpy_network_providers.resources import EmptyAddress
+from erdpy_network_providers.transaction_events import TransactionEvent
 
 
 class TransactionLogs:
